add support for floats in dsl and parser

+++ app/HotDrink.hs
@@ -3,7 +3,7 @@ module HotDrink where
 import           MethodParser
 
 type Identifier = String
-type Variable = (Identifier, Maybe Int)
+type Variable = (Identifier, Maybe Double)
 type Method = (Identifier, Identifier, Expr)
 
 data VertexType
